Add rpm-py-installer as a dependency for the integration tests
'rpm' is an optional dependency for kobo, which is required by the
integration tests.

Previously this dependency was implicitly satisfied by 'koji' pulling
in 'rpm-py-installer' which made sure 'rpm' is available. But
koji==1.20.0 made this dependency optional, so importing kobo.rpmlib
started failing.

Fix this by making 'rpm-py-installer' and explicit dependency for the
integration test environment.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>